Start Switching to External NuGets for Native Build
Removes the explicit `<link>` references and adds a reference to the
Mac NuGet build. This only exists locally for the time being. The plan
is to have one of these 'battery' NuGet packages for each supported
native platform. The battery packages make use of the
`runtimes/native` folder structure to include the native assets for
each target runtime.