Add Invoke Callback Support
This seems to work. We keep hold of the callback the user has given us
and register our own in its place. When an invoke request arrives back
validate it really is for this instance and forward the request on to
the user's action.

NOTE: This doesn't properly handle the pinning of the delegate here. Not
sure what the full solution is to that yet.